+++ src/containers/AuthenticationPage/ConfirmSignupForm/ConfirmSignupForm.js
@@ -1,11 +1,12 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { compose } from 'redux';
-import { FormattedMessage, injectIntl, intlShape } from '../../util/reactIntl';
 import { Form as FinalForm } from 'react-final-form';
 import classNames from 'classnames';
-import * as validators from '../../util/validators';
-import { Form, PrimaryButton, FieldTextInput } from '../../components';
+
+import { FormattedMessage, injectIntl, intlShape } from '../../../util/reactIntl';
+import * as validators from '../../../util/validators';
+import { Form, PrimaryButton, FieldTextInput } from '../../../components';
 
 import css from './ConfirmSignupForm.module.css';