+++ src/containers/AuthenticationPage/LoginForm/LoginForm.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import { renderDeep } from '../../util/test-helpers';
-import { fakeIntl } from '../../util/test-data';
+import { renderDeep } from '../../../util/test-helpers';
+import { fakeIntl } from '../../../util/test-data';
 import LoginForm from './LoginForm';
 
 const noop = () => null;